Add files via upload
Adding a box in each page to explain what I want to be on each page before I meet with my tutor so he isn't looking at a blank screen and has a feel for the layout I want for the web app. Also added a feature to the logo so that if it is clicked then it will send the user back to the page of contents.