Add files via upload
I have changed the landing page for the web app so now when the user first goes on to the app they will be met with an input box and a landing page where they will have to input their specific tracker code into the box and then when the button is clicked then it will send them to the contents page. For now, when the button on the landing page is clicked then it will send the user straight to the contents page. However, for the end product, I want to make it so that if the user inputs an incorrect tracker code then it will not allow them to go further and an error message will pop up to the user.