Add files via upload
Began to change the look of some of the pages after receiving feedback about my work. I have also added in a Loading screen that flashes up before the user gets to the contents page. It shows up for a few seconds and then sends the user straight to the contents page