Add files via upload
Added a loading screen for each of the pages when the specific buttons have been clicked. For example, when they want to go on the heart rate page it will send them to the loading screen and then to the heart rate screen. Also added information on each of the stats pages to make them different to each other