Add files via upload
Changed the sizing and positioning of all of the elements so they are now all responsive to when the page is made smaller or larger by the user. Also added loaders between each page to make it look a bit more professional. I also changed the layout of the contents page so everything was more spread out and not so clumped together